Story_Composer/requirement.txt
fpdf==1.7.2
Pillow==10.2.0